Moved module producttype back to nckcindex (caused circular dependency); Import fixes.
