+++ xcube_smos/nckcindex/producttype.py
